fix: pin binary byte order and Cython integer widths
Two portability defects found while auditing platform independence.
Both are latent on Linux x86-64 — neither changes behaviour there —
but both are genuine correctness bugs, and the first is an
inconsistency worth removing regardless of which platforms we
support.

1. Spots `.pos` reader used native byte order
-------------------------------------------------------------------
`sync_pipelines` read the spots tracker's `.pos` file with a native
`np.int16`.  It was the only raw-binary read in the package that did
not pin byte order — the other twelve all use explicit `<i2`, `<i4`,
`<i8`, `<f4`, `<f8`, or a `np.dtype("<i{n}")` constant.  On a
big-endian host it would silently yield byte-swapped coordinates
rather than failing.  Now `"<i2"`.

`sync_nlx_spots` has no test coverage, and a behavioural test could
not have caught this anyway: on little-endian hardware `np.int16` and
`"<i2"` are byte-for-byte identical, so no assertion running on x86 or
Apple Silicon can distinguish them.  Added
`tests/test_binary_byte_order.py` instead — a source-level guard that
parses every module under `neurobox/{io,dtype,analysis,viz,config,
utils}` and asserts each `np.fromfile` / `np.frombuffer` / `np.memmap`
call names an explicit byte order, including dtypes hoisted into
module constants and built via f-strings.  Single-byte and
order-irrelevant specs (`i1`, `u1`, `S`, `V`, `?`) are accepted.

The guard is self-tested (it must reject `np.int16` and accept `<i2`,
`>f8`, `|i1`, `u1`) and was verified against the real defect: with the
fix reverted it fails and names `sync_pipelines.py:800` exactly.

2. Cython kernels used bare C `long` for int64 buffers
-------------------------------------------------------------------
`_ccg_engine.pyx` and `_within_ranges_engine.pyx` declared their
integer buffers as `cnp.ndarray[long, ...]` while allocating the
backing arrays with `dtype=np.int64` and receiving `np.int64` from
every Python caller.  The two coincide under LP64 (Linux, macOS), but
C `long` is 32 bits under Windows' LLP64 model, so there the declared
buffer type and the actual array dtype disagree and every CCG /
within-ranges call raises `ValueError: Buffer dtype mismatch`.
Confirmed by compiling both widths and feeding each the arrays
`ccg.py` builds.

Retyped to fixed-width `cnp.int64_t` throughout — buffer
declarations, scalars, pointers, memoryviews and `<long>` casts — so
the declared type and the allocated dtype stay in lockstep under any
data model.  A short note in each module docstring records the reason,
to stop it being "simplified" back to `long`.

Verified as a pure no-op on LP64.  A golden-reference harness hashes
the exact output bytes and dtypes of both kernels across 80
randomised trials (varying spike counts, group counts, bin sizes and
half-bin widths); the digest is identical before and after the change
on CPython 3.10.20, 3.12.3, 3.13.13 and 3.14.4.

Not addressed, by decision: Windows would additionally need a
fallback for the unprivileged `symlink_to` in `link_session` and a
non-POSIX `data_root` default.  Linux remains the only supported
platform and the `Operating System :: POSIX :: Linux` classifier
stays accurate; these two fixes are about correctness and internal
consistency, not about adding platform support.

Suite: 1090 passed / 6 skipped on 3.10, 3.12, 3.13 and 3.14
(1083 + 7 new byte-order guard tests).

+++ neurobox/dtype/sync_pipelines.py
@@ -793,7 +793,11 @@ def sync_nlx_spots(
             f"Spots .pos file not found for {session.name!r}.  "
             f"Searched: {session.spath}"
         )
-    raw  = np.fromfile(str(pos_file), dtype=np.int16).reshape(-1, 4)
+    # Explicit little-endian: the .pos format is LE on disk, and every
+    # other binary reader in neurobox pins byte order the same way.
+    # A native `np.int16` here would silently byte-swap on a big-endian
+    # host rather than failing loudly.
+    raw  = np.fromfile(str(pos_file), dtype="<i2").reshape(-1, 4)
     xy   = raw.astype(np.float64).reshape(-1, 2, 2)
     n_fr = xy.shape[0]
     xyz_arr = np.concatenate([xy, np.zeros((n_fr, 2, 1))], axis=-1)
